+++ api/src/main/java/mb/resource/hierarchical/walk/TrueResourceWalker.java
@@ -17,6 +17,6 @@ public class TrueResourceWalker implements ResourceWalker {
     }
 
     @Override public String toString() {
-        return "TrueResourceWalker()";
+        return "true";
     }
 }